==> scripts/markdown/01_patient_characteristics.Rmd <==
---
title: "Impact of dysfunctional ubiquitination in response to cancer immunotherapy"
subtitle: <center> Organizing patients characteristics </center>
author: "Eric Bautista Farrerons (s212514)"
date: '`r paste("First created on January 2024. Updated on ", format(Sys.Date(), "%d %B %Y"))`'
output:
  html_document:
    css: style.css
    code_folding: hide
    fig_caption: yes
    toc: yes
    toc_float:
      collapsed: true
      smooth_scroll: true
    toc_depth: 4
    number_sections: true
---

```{r, message=FALSE, warning=FALSE}
### Load packages, functions and paths
source(file = "../01_install.R")
source(file = "../02_functions.R")
source(file = "../03_paths.R")

# Sup1
sup1_data <- "41467_2017_1460_MOESM4_ESM_clinical.xlsx"
sup1_path <- file.path(data_path,
                       sup1_data)

# RData obj
mm909_data <- "MM909_data.RData"
mm909_path <- file.path(data_path,
                        mm909_data)

# Provean
provean_path <- "../../data/provean_output/provean_output"

# MAF-like file
maf_path <- file.path(data_path,
                      "41467_2017_1460_MOESM6_ESM_somatic_mutations.xlsx")

### Read data
sup1_table <- read_excel(sup1_path,
                         skip = 1,
                         col_names = TRUE)
colnames(sup1_table)[1] <- "Sample.ID"

load(mm909_path)

maf_table <- read_excel(maf_path,
                        skip = 1,
                        col_names = TRUE)

### Clean data
# Response group definition
sup1_response <- sup1_table %>%
  mutate(Patient.Response = case_when(
    RECIST %in% c("CR", "PR") ~ "R",
    RECIST %in% c("SD") & OS.Time >= 24 ~ "R",
    RECIST %in% c("SD") & OS.Time < 24 ~ "NR",
    RECIST %in% c("PD") ~ "NR",
    TRUE ~ NA_character_  # For any cases that don't match above conditions
  )
  )

sup1_final <- sup1_response %>%
  left_join(response %>%
      dplyr::select(patient, mut_load),
    by = c("Sample.ID" = "patient")
  )

# Subset sup1_table
sup1_subset <- sup1_final %>%
  dplyr::select(Sample.ID,
                Sample.Timepoint,
                AJCC.Stage,
                PFS.Time,
                PFS.Event,
                OS.Time,
                OS.Event,
                Type.of.Lesion,
                Type.of.Primary,
                RECIST,
                Patient.Response,
                mut_load)

# Save the sup1_subset into a file
sup1_out_path <- file.path(results_pc_path,
                           "sup1_response.csv")
write.csv(sup1_subset, file = sup1_out_path, row.names = TRUE)

# Save the sup1_final into a file
clinical_out_path <- file.path(results_pc_path,
                           "clinical_metadata.csv")
write.csv(sup1_final, file = clinical_out_path, row.names = TRUE)
```

## Patients' characteristics with response to treatment

```{r}
datatable(sup1_final, 
          extensions = c('Buttons', 
                         'FixedColumns'), 
          options = list(
            dom = 'Bfrtip',
            buttons = c('copy', 'excel', 'csv'),
            scrollX=TRUE,
            pageLength=10,
            fixedColumns = list(leftColumns = 2), # Freeze the first 2 columns
            columnDefs = list(list(
              targets = "_all",
              render = JS(
                "function(data, type, row, meta) {",
                "  return data === null ? 'NA' : data;",
                "}"
              )
            ))
          ),
          caption = "Patients' characteristics with response to treatment & mut_load"
        )
```

## Group by response

```{r}
response_df <- sup1_final %>% 
  dplyr::count(Patient.Response)

datatable(response_df, 
          extensions = 'Buttons', 
          options = list(
            dom = 'Bfrtip',
            buttons = c('copy', 'excel', 'csv'),
            scrollX=TRUE,
            pageLength=10,
            columnDefs = list(list(
              targets = "_all",
              render = JS(
                "function(data, type, row, meta) {",
                "  return data === null ? 'NA' : data;",
                "}"
              )
            ))
          ),
          caption = "Patients' grouped by response"
        )
```

## Extract patients' IDs removed from RData object

```{r}
### Extract rows with NAs in mut_load
rows_with_na <- sup1_final[is.na(sup1_final$mut_load), ]
datatable(rows_with_na, 
          extensions = c('Buttons', 
                         'FixedColumns'), 
          options = list(
            dom = 'Bfrtip',
            buttons = c('copy', 'excel', 'csv'),
            scrollX=TRUE,
            pageLength=5,
            fixedColumns = list(leftColumns = 2), # Freeze the first 2 columns
            columnDefs = list(list(
              targets = "_all",
              render = JS(
                "function(data, type, row, meta) {",
                "  return data === null ? 'NA' : data;",
                "}"
              )
            ))
          ),
          caption = "Patients' removed from RData object."
        )
```

## Create a list of patients' IDs present in MAF file

```{r}
### Sort IDs from maf_table
sorted_ids <- maf_table %>%
  dplyr::select(tumor_name) %>%
  distinct(tumor_name) %>%
  arrange(tumor_name)

# Display the result
datatable(sorted_ids, 
          extensions = 'Buttons', 
          options = list(
            dom = 'Bfrtip',
            buttons = c('copy', 'excel', 'csv'),
            scrollX=TRUE,
            pageLength=24,
            columnDefs = list(list(
              targets = "_all",
              render = JS(
                "function(data, type, row, meta) {",
                "  return data === null ? 'NA' : data;",
                "}"
              )
            ))
          ),
          caption = "Sorted IDs from maf_table"
        )
```

## Check if patients' IDs excluded in RData are present in MAF file

```{r}
search_ids <- c("MM909_01", "MM909_20", "MM909_24", "MM909_29", "MM909_45")

# Check if each ID is in the unique_sorted_ids dataframe
is_id_present <- sapply(search_ids, 
                        function(id) any(str_detect(sorted_ids$tumor_name, id)))

# Combine the IDs and their presence status into a data frame for easy viewing
patients_present <- data.frame(Sample.ID = search_ids, Is.Present = is_id_present)

# Display the results
datatable(patients_present, 
          extensions = 'Buttons', 
          options = list(
            dom = 'Bfrtip',
            buttons = c('copy', 'excel', 'csv'),
            scrollX=TRUE,
            pageLength=24,
            columnDefs = list(list(
              targets = "_all",
              render = JS(
                "function(data, type, row, meta) {",
                "  return data === null ? 'NA' : data;",
                "}"
              )
            ))
          ),
          caption = "Patient IDs excluded in RData present/absent in MAF file"
        )
```

## Conclusions

1. There are 3 patients that are absent in the RData and also in the MAF-like file:

  - MM909_01
  - MM909_29
  - MM909_45

  This is because their samples failed the WES. It is described in the paper.

2. There are 2 patients that are absent in the RData but present in the MAF-like file:

  - MM909_20
  - MM909_24

  This is because they don't have RNA-Seq data. It is described in the paper.
  
## Results table

```{r}
# Merge the 2 dataframes on "Sample.ID"
tmp_merge <- left_join(sup1_final, patients_present, by = "Sample.ID")

results_df <- tmp_merge %>%
  mutate(RNASeq.Data = case_when(
    Is.Present == TRUE ~ "NO",
    Is.Present == FALSE | is.na(Is.Present) ~ "YES")
    ) %>% 
  mutate(WES.Data = case_when(
    Is.Present == TRUE | is.na(Is.Present) ~ "YES",
    Is.Present == FALSE ~ "FAILED")
    ) %>% 
  dplyr::select(Sample.ID,
                RNASeq.Data,
                WES.Data)

# Display the results
datatable(results_df, 
          extensions = 'Buttons', 
          options = list(
            dom = 'Bfrtip',
            buttons = c('copy', 'excel', 'csv'),
            scrollX=TRUE,
            pageLength=10,
            columnDefs = list(list(
              targets = "_all",
              render = JS(
                "function(data, type, row, meta) {",
                "  return data === null ? 'NA' : data;",
                "}"
              )
            ))
          ),
          caption = "Present/missing data per patient"
        )
```
  
# Final clinical_df table

```{r}
final_clincal_data <- sup1_final %>% 
  dplyr::left_join(results_df, by="Sample.ID")

saveRDS(final_clincal_data, file = file.path(results_pc_path, "extended_clinical_data.rds"))
```

## Characteristics plot

```{r, message=FALSE, warning=FALSE, fig.width=8, fig.height=6}
# Fig1
pat_df <- sup1_final %>%
  dplyr::rename(patient_response = Patient.Response,
                sample_id = Sample.ID,
                PrimaryMelanoma = Type.of.Primary) %>%
  dplyr::mutate(os_binary = ifelse(OS.Time >= 24, "1", "0")) %>% 
  dplyr::arrange(mut_load)

  cbPalette <- c("#56B4E9", "#999999", "#8A2BE2", "#E69F00", "#00CED1", "#CC79A7", "#4B0082", "#F0E442", "#0072B2", "#FF69B4", "#000000")

p1 <- ggplot(pat_df, aes(x = reorder(sample_id, mut_load), y = mut_load, fill = patient_response)) +
  geom_bar(stat = "identity", width = 0.7, linewidth = 0.5, color = "black") +
  scale_fill_manual(values = c(R = "#009E73", NR = "#D55E00"),
                    labels = c("R" = "Responders", "NR" = "Non-responders")) +
  labs(x = NULL, 
       y = "Mutation Load",
       fill = "Patient Response",
       title = "Mutational load and clinical features of the cohort") +
  theme_minimal() +
  theme(axis.text.x = element_blank(), 
        axis.ticks.x = element_blank(),
        legend.position = "right")

p2 <- make_annotation_plot(pat_df, "RECIST", "RECIST", 1, c("PD" = "PD", "SD" = "SD", "PR" = "PR", "CR" = "CR"))
p3 <- make_annotation_plot(pat_df, "os_binary", "Overall survival (OS)", 5, c("1" = "OS >= 24", "0" = "OS < 24"))
p4 <- make_annotation_plot(pat_df, "PrimaryMelanoma", "Type of primary melanoma", 7, c("mucosa" = "Mucosal", "skin" = "Cutaneous", "unknown" = "Unknown"), show_sample_id = TRUE)

# Extract legends
legends <- list(
  p1_legend = get_legend(p1),
  p2_legend = get_legend(p2),
  p3_legend = get_legend(p3),
  p4_legend = get_legend(p4)
)

# Combine plots with legends
combined_plot <- (p1 + theme(legend.position = "none")) /
                 (p2 + theme(legend.position = "none")) /
                 (p3 + theme(legend.position = "none")) /
                 (p4 + theme(legend.position = "none")) /
  plot_layout(nrow=4, ncol=1)

combined_legends <- (as_ggplot(legends$p1_legend)) +
                    (as_ggplot(legends$p2_legend)) +
                    (as_ggplot(legends$p3_legend)) +
                    (as_ggplot(legends$p4_legend)) +
  plot_layout(nrow=1, ncol=4)

final_plot <- combined_plot / combined_legends +
  plot_layout(nrow=5, ncol=1, heights=c(4, 0.3, 0.3, 0.3, 2))
  
# Save and print the plot
plot_path <- file.path(figures_pc_path, 
                      "pat_char_summary.png")
ggsave(plot_path, final_plot, width = 8, height = 6, dpi = 300)
print(final_plot)
```

## Session Info
  
```{r}
sessionInfo()
```
